+++ src/modules/logger.service.ts
@@ -67,9 +67,6 @@ export class LoggerService {
    */
   public error(message: string, error?: any): void {
     console.error(chalk.red(message));
-    if (error) {
-      
-    }
   }
 
     /**